+++ canvas_modules/common-canvas/src/common-canvas/svg-canvas-utils-external.js
